requirements.txt
streamlit
pandas
qrcode
Pillow
gspread
oauth2client
gspread-dataframe
google-auth